+++ aurora-springboot/src/main/java/com/aurora/mapper/UserAuthMapper.java
@@ -15,6 +15,6 @@ public interface UserAuthMapper extends BaseMapper<UserAuth> {
 
     List<UserAdminDTO> listUsers(@Param("current") Long current, @Param("size") Long size, @Param("conditionVO") ConditionVO conditionVO);
 
-    Integer countUser(@Param("conditionVO") ConditionVO conditionVO);
+    Long countUser(@Param("conditionVO") ConditionVO conditionVO);
 
 }
