+++ aurora-springboot/src/main/java/com/aurora/service/impl/TalkServiceImpl.java
@@ -41,7 +41,7 @@ public class TalkServiceImpl extends ServiceImpl<TalkMapper, Talk> implements Ta
 
     @Override
     public PageResultDTO<TalkDTO> listTalks() {
-        Integer count = talkMapper.selectCount((new LambdaQueryWrapper<Talk>()
+        Long count = talkMapper.selectCount((new LambdaQueryWrapper<Talk>()
                 .eq(Talk::getStatus, PUBLIC.getStatus())));
         if (count == 0) {
             return new PageResultDTO<>();
@@ -92,7 +92,7 @@ public class TalkServiceImpl extends ServiceImpl<TalkMapper, Talk> implements Ta
 
     @Override
     public PageResultDTO<TalkAdminDTO> listBackTalks(ConditionVO conditionVO) {
-        Integer count = talkMapper.selectCount(new LambdaQueryWrapper<Talk>()
+        Long count = talkMapper.selectCount(new LambdaQueryWrapper<Talk>()
                 .eq(Objects.nonNull(conditionVO.getStatus()), Talk::getStatus, conditionVO.getStatus()));
         if (count == 0) {
             return new PageResultDTO<>();
